Have a plane always non-zero
Without a non-zero plane, no way for network to tell if zeros are from
padding or zeros are information. Network cannot be translation invariant,
so it needs an always non-zero plane. I didn't want to add another plane,
because we already have 8 (power of 2). Instead, the mover plane can be
set to be either +1 or -1 depending on mover, which should help the network
determine padding vs not with a bit of extra logic. Swapped mover order so
that X corresponds to 1 and O to -1.